refactor: Rename helpers in makeTheme and extract all helpers
in specific files
