+++ react/MuiCozyTheme/makeOverrides.js
@@ -1,210 +1,12 @@
-import React from 'react'
-import { createMuiTheme } from '@material-ui/core/styles'
-
-import {
-  getCssVariableValue,
-  getInvertedCssVariableValue
-} from '../utils/color'
-import isTesting from '../helpers/isTesting'
-import AccordionExpandIcon from './AccordionExpandIcon'
+import { getCssVariableValue } from '../utils/color'
 
 const SWITCH_BAR_WIDTH = 25
 
-const themesCommonConfig = {
-  shape: {
-    borderRadius: 6
-  },
-  breakpoints: {
-    values: {
-      xs: 0,
-      sm: 480,
-      md: 768,
-      lg: 1023,
-      xl: 1200
-    }
-  },
-  zIndex: {
-    modal: getCssVariableValue('zIndex-modal')
-  },
-  props: {
-    MuiTabs: {
-      textColor: 'primary',
-      TabIndicatorProps: { color: 'primary' }
-    },
-    MuiButton: {
-      disableRipple: true
-    },
-    MuiListItem: {
-      disableRipple: true
-    },
-    MuiTooltip: {
-      arrow: true
-    },
-    MuiAccordionSummary: {
-      expandIcon: <AccordionExpandIcon />
-    }
-  },
-  ...(isTesting() && { transitions: { create: () => 'none' } })
-}
+export const makeThemeOverrides = theme => {
+  const createOverrides =
+    theme.palette.type === 'dark' ? makeInvertedOverrides : makeOverrides
 
-const makeTypography = palette => ({
-  fontFamily: getCssVariableValue('primaryFont') || 'Lato',
-  h1: {
-    fontSize: 48,
-    color: palette.text.primary,
-    fontWeight: 'bold',
-    lineHeight: 1.087,
-    letterSpacing: -0.8
-  },
-  h2: {
-    fontSize: 32,
-    color: palette.text.primary,
-    fontWeight: 'bold',
-    lineHeight: 1.313,
-    letterSpacing: -0.4
-  },
-  h3: {
-    fontSize: 24,
-    color: palette.text.primary,
-    fontWeight: 'bold',
-    lineHeight: 1.167
-  },
-  h4: {
-    fontSize: 20,
-    color: palette.text.primary,
-    fontWeight: 'bold',
-    lineHeight: 1.167
-  },
-  h5: {
-    fontSize: 18,
-    color: palette.text.primary,
-    fontWeight: 'bold',
-    lineHeight: 1.313
-  },
-  h6: {
-    fontSize: 16,
-    color: palette.text.primary,
-    fontWeight: 'bold',
-    lineHeight: 1.313
-  },
-  subtitle1: {
-    fontWeight: 'bold',
-    fontSize: 12,
-    lineHeight: 1.313,
-    textTransform: 'uppercase'
-  },
-  body1: {
-    fontSize: 16,
-    lineHeight: 1.313,
-    color: palette.text.primary
-  },
-  body2: {
-    fontSize: 14,
-    lineHeight: 1.313,
-    color: palette.text.primary
-  },
-  button: {
-    fontWeight: 'bold',
-    fontSize: 14,
-    lineHeight: 1.313
-  },
-  caption: {
-    fontSize: 12,
-    lineHeight: 1.313,
-    color: palette.text.secondary,
-    display: 'block'
-  }
-})
-
-const makePalette = type => {
-  const getCssValue =
-    type === 'dark' ? getInvertedCssVariableValue : getCssVariableValue
-
-  return {
-    type,
-    primary: {
-      light: getCssValue('primaryColorLight'),
-      main: getCssValue('primaryColor'),
-      dark: getCssValue('primaryColorDark'),
-      contrastText: getCssValue('primaryContrastTextColor')
-    },
-    secondary: {
-      light: getCssValue('secondaryColorLight'),
-      main: getCssValue('secondaryColor'),
-      dark: getCssValue('secondaryColorDark'),
-      contrastText: getCssValue('secondaryContrastTextColor')
-    },
-    error: {
-      light: getCssValue('errorColorLight'),
-      main: getCssValue('errorColor'),
-      dark: getCssValue('errorColorDark'),
-      contrastText: getCssValue('errorColorContrastText')
-    },
-    warning: {
-      light: getCssValue('warningColorLight'),
-      main: getCssValue('warningColor'),
-      dark: getCssValue('warningColorDark'),
-      contrastText: getCssValue('warningColorContrastText')
-    },
-    success: {
-      light: getCssValue('successColorLight'),
-      main: getCssValue('successColor'),
-      dark: getCssValue('successColorDark'),
-      contrastText: getCssValue('successColorContrastText')
-    },
-    info: {
-      light: getCssValue('infoColorLight'),
-      main: getCssValue('infoColor'),
-      dark: getCssValue('infoColorDark'),
-      contrastText: getCssValue('infoColorContrastText')
-    },
-    text: {
-      primary: getCssValue('primaryTextColor'),
-      secondary: getCssValue('secondaryTextColor'),
-      disabled: getCssValue('disabledTextColor'),
-      hint: getCssValue('hintTextColor'),
-      icon: getCssValue('iconTextColor')
-    },
-    grey: {
-      0: getCssValue('grey0'),
-      50: getCssValue('grey50'),
-      100: getCssValue('grey100'),
-      200: getCssValue('grey200'),
-      300: getCssValue('grey300'),
-      400: getCssValue('grey400'),
-      500: getCssValue('grey500'),
-      600: getCssValue('grey600'),
-      700: getCssValue('grey700'),
-      800: getCssValue('grey800'),
-      900: getCssValue('grey900'),
-      A100: getCssValue('greyA100'),
-      A200: getCssValue('greyA200'),
-      A400: getCssValue('greyA400'),
-      A700: getCssValue('greyA700')
-    },
-    divider: getCssValue('dividerColor'),
-    action: {
-      active: getCssValue('actionColorActive'),
-      hover: getCssValue('actionColorHover'),
-      selected: getCssValue('actionColorSelected'),
-      disabled: getCssValue('actionColorDisabled'),
-      disabledBackground: getCssValue('actionColorDisabledBackground'),
-      focus: getCssValue('actionColorFocus'),
-      ghost: getCssValue('actionColorGhost'),
-      hoverGhost: getCssValue('actionColorHoverGhost')
-    },
-    border: {
-      main: getCssValue('borderMainColor'),
-      disabled: getCssValue('borderDisabledColor'),
-      ghost: getCssValue('borderGhostColor'),
-      ghostDisabled: getCssValue('borderGhostDisabledColor')
-    },
-    background: {
-      default: getCssValue('defaultBackgroundColor'),
-      paper: getCssValue('paperBackgroundColor'),
-      selected: '#F5FAFF'
-    }
-  }
+  return createOverrides(theme)
 }
 
 const makeOverrides = theme => ({
@@ -766,55 +568,4 @@ const makeInvertedOverrides = invertedTheme => {
   }
 
   return invertedOverrides
-}
-
-const makeShadows = () => [
-  'none',
-  'rgba(29, 33, 42, 0.08) 0px 2px 4px 0px, rgba(29, 33, 42, 0.06) 0px 4px 16px 0px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.08) 0px 3px 5px 0px, rgba(29, 33, 42, 0.06) 0px 4px 17px 0px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.09) 0px 4px 6px 0px, rgba(29, 33, 42, 0.07) 0px 5px 19px 1px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.09) 0px 5px 8px 1px, rgba(29, 33, 42, 0.07) 0px 5px 20px 1px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.09) 0px 6px 9px 1px, rgba(29, 33, 42, 0.07) 0px 5px 22px 1px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.1) 0px 7px 10px 1px, rgba(29, 33, 42, 0.07) 0px 6px 23px 2px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.1) 0px 8px 11px 1px, rgba(29, 33, 42, 0.08) 0px 6px 24px 2px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.1) 0px 9px 13px 1px, rgba(29, 33, 42, 0.08) 0px 6px 26px 2px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.11) 0px 10px 14px 1px, rgba(29, 33, 42, 0.08) 0px 7px 27px 3px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.11) 0px 11px 15px 2px, rgba(29, 33, 42, 0.08) 0px 7px 29px 3px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.11) 0px 12px 16px 2px, rgba(29, 33, 42, 0.09) 0px 7px 30px 3px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.12) 0px 13px 17px 2px, rgba(29, 33, 42, 0.09) 0px 8px 31px 4px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.12) 0px 13px 19px 2px, rgba(29, 33, 42, 0.09) 0px 8px 33px 4px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.13) 0px 14px 20px 2px, rgba(29, 33, 42, 0.09) 0px 9px 34px 5px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.13) 0px 15px 21px 2px, rgba(29, 33, 42, 0.1) 0px 9px 35px 5px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.13) 0px 16px 22px 3px, rgba(29, 33, 42, 0.1) 0px 9px 37px 5px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.14) 0px 17px 23px 3px, rgba(29, 33, 42, 0.1) 0px 10px 38px 6px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.14) 0px 18px 25px 3px, rgba(29, 33, 42, 0.1) 0px 10px 40px 6px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.14) 0px 19px 26px 3px, rgba(29, 33, 42, 0.11) 0px 10px 41px 6px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.15) 0px 20px 27px 3px, rgba(29, 33, 42, 0.11) 0px 11px 42px 7px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.15) 0px 21px 28px 3px, rgba(29, 33, 42, 0.11) 0px 11px 44px 7px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.15) 0px 22px 30px 4px, rgba(29, 33, 42, 0.11) 0px 11px 45px 7px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.16) 0px 23px 31px 4px, rgba(29, 33, 42, 0.12) 0px 12px 47px 8px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px',
-  'rgba(29, 33, 42, 0.16) 0px 24px 32px 4px, rgba(29, 33, 42, 0.12) 0px 12px 48px 8px, rgba(29, 33, 42, 0.12) 0px 0px 0px 0.5px'
-]
-
-const makeThemeOverrides = theme => {
-  const createOverrides =
-    theme.palette.type === 'dark' ? makeInvertedOverrides : makeOverrides
-
-  return createOverrides(theme)
-}
-
-export const makeTheme = type => {
-  const palette = makePalette(type)
-  const theme = createMuiTheme({
-    ...themesCommonConfig,
-    typography: makeTypography(palette),
-    palette,
-    shadows: makeShadows()
-  })
-  const overrides = makeThemeOverrides(theme)
-
-  return {
-    ...theme,
-    overrides
-  }
 }
