+++ react/MuiCozyTheme/makeTypography.js
@@ -1,0 +1,70 @@
+import { getCssVariableValue } from '../utils/color'
+
+export const makeTypography = palette => ({
+  fontFamily: getCssVariableValue('primaryFont') || 'Lato',
+  h1: {
+    fontSize: 48,
+    color: palette.text.primary,
+    fontWeight: 'bold',
+    lineHeight: 1.087,
+    letterSpacing: -0.8
+  },
+  h2: {
+    fontSize: 32,
+    color: palette.text.primary,
+    fontWeight: 'bold',
+    lineHeight: 1.313,
+    letterSpacing: -0.4
+  },
+  h3: {
+    fontSize: 24,
+    color: palette.text.primary,
+    fontWeight: 'bold',
+    lineHeight: 1.167
+  },
+  h4: {
+    fontSize: 20,
+    color: palette.text.primary,
+    fontWeight: 'bold',
+    lineHeight: 1.167
+  },
+  h5: {
+    fontSize: 18,
+    color: palette.text.primary,
+    fontWeight: 'bold',
+    lineHeight: 1.313
+  },
+  h6: {
+    fontSize: 16,
+    color: palette.text.primary,
+    fontWeight: 'bold',
+    lineHeight: 1.313
+  },
+  subtitle1: {
+    fontWeight: 'bold',
+    fontSize: 12,
+    lineHeight: 1.313,
+    textTransform: 'uppercase'
+  },
+  body1: {
+    fontSize: 16,
+    lineHeight: 1.313,
+    color: palette.text.primary
+  },
+  body2: {
+    fontSize: 14,
+    lineHeight: 1.313,
+    color: palette.text.primary
+  },
+  button: {
+    fontWeight: 'bold',
+    fontSize: 14,
+    lineHeight: 1.313
+  },
+  caption: {
+    fontSize: 12,
+    lineHeight: 1.313,
+    color: palette.text.secondary,
+    display: 'block'
+  }
+})
